new Parameter @CleanUpStartTime / @CleanUpEndTime
When @CleanUpStartTime and @CleanUpEndTime are specified the delition of old backup files will be only executed, when the procedure was called in this timeframe. 

Reason: xp_deletefile can be very slow, if there are many files in the backup directory (e.g. because LOG backups are taken every minute and cleaned up after 14 days -> 14 * 24 * 60 = 20160 files) and should for this reason only been executed once per night to prevent delays in the main working time.