Enhanced Mirroring
You can now specify up to 3 mirror destinations (before only one), each with separat cleanup time and cleanup mode (no backup to URL support for this and only 2 mirrors when you are using idera Backup as backup tool)

Instead of using the build in MIRROR parameter (Enterprise feature) you can set @MirrorType to 'COPY' or 'COPY_LATER' to use xp_cmd_shell to invoke ROBOCOPY, XCOPY or COPY to copy the created backup files to the mirror directory / directories. 'COPY' is automatical choosen, if you specify a @MirrorDirectory but have no Enterprise / Developer Edition.

The new parameters @MirrorWhenNotAvailable to @MirrorWhenNotAvailable3 lets you specify, what to do if the mirror destination is not available (temporary network problem, reboot ...) - 'ERROR' (default), 'SKIP' or 'COPY_LATER' (most useful).

When either one of the  @MirrorType or @MirrorWhenNotAvailable parameters is set to 'COPY_LATER' it creates entries with CommandType LIKE 'PENDING%' in the master.CommandLog table. At the next execution it will process this entries (except @SkipPendingMirrorCopies is set to 'Y') and copies / verifies the pending files to the mirrors (if they are available and the @MirrorCleanupTime is not reached). 

The copy-later-feature is particularly useful, if the mirror destination is on a remote site with a slower connection and you do not want to delay your regular every-5-minute-log-backup, just because it takes 10 minutes to transfer this big log backup after your index optimize run. To make full use of it, you should set @MirrorType to 'COPY_LATER', @SkipPendingMirrorCopies to 'Y' and create a second SQL Agent job which calls dbo.DatabaseBackup with @BackupType = 'COPY_PENDING_MIRRORS' (this will skip all the other backup stuff)